[REF] website_sale: convert smaller Public Widgets to Interactions
task-4919046
